==> notebooks/1_data.ipynb <==
# Downloading data from Quesmed

Direct DB connection to download the latest comments

import psycopg2
from dotenv import load_dotenv
import os

load_dotenv()

conn = psycopg2.connect(
  dbname=os.getenv("QUESMED_POSTGRES_DB"),
  user=os.getenv("QUESMED_POSTGRES_USER"),
  password=os.getenv("QUESMED_POSTGRES_PASSWORD"),
  port=os.getenv("QUESMED_POSTGRES_PORT"),
  host=os.getenv("QUESMED_POSTGRES_HOST")
)

keys = (
  'id', 
  'createdAt', 
  'userId', 
  'userCreatedAt', 
  'classYear', 
  'universityId',
  'country',
  'universityName',
  'parentId',
  'questionId',
  'comment',
  'review',
  'likes',
  'dislikes'
)
data_map = {k: [] for k in keys}
data_map

import pandas as pd
from tqdm import tqdm

file_path = "../data/1comments.h5"

def load_df():
  if os.path.isfile(file_path):
    df = pd.read_hdf(file_path, key='df')
    return df

  cur = conn.cursor()
  cur.execute("""
SELECT
  com.id,
  com."createdAt",
  com."userId",
  u."createdAt" "userCreatedAt",
  u."classYear",
  u."universityId",
  uni.country,
  uni.name "universityName",
  com."parentId",
  com."questionId",
  com.comment,
  com.review,
  sum(qcl."likeTrueDislikeFalse"::int) "likes",
  sum(abs(qcl."likeTrueDislikeFalse"::int - 1)) "dislikes"
FROM
  question_comments com
  INNER JOIN users u ON com."userId" = u.id
  INNER JOIN universities uni ON u."universityId" = uni.id
  LEFT JOIN question_comment_likes qcl ON com.id = qcl."commentId"
    AND qcl."likeTrueDislikeFalse" IS NOT NULL
GROUP BY
  com.id,
  u.id,
  uni.id
ORDER BY
  com."createdAt" ASC;
  """)

  t = tqdm()
  batch_size = 1000
  loading = True
  while loading:
      data = cur.fetchmany(batch_size)
      t.update(len(data))
      if len(data) < batch_size:
          loading = False
      for row in data:
          for i, k in enumerate(keys):
              data_map[k].append(row[i])
  t.close()
  cur.close()

  df = pd.DataFrame.from_dict(data_map)
  df.to_hdf(file_path, key='df', mode='w')
  return df

df = load_df()
print(df.shape)
df.sample(3)
