==> notebooks/2_cleaning.ipynb <==
import pandas as pd

file_path = "../data/1comments.h5"
df = pd.read_hdf(file_path, key='df')
df.sample(3)

## Processing DataFrame

f"Date range of comments: {df.iat[0, 1]} to {df.iat[-1, 1]}"

blank_comments = df.comment.str.len() > 2
blank_comments.value_counts()

df = df.loc[blank_comments]
df.shape

df['likes'] = df['likes'].fillna(0).astype(int)
df['dislikes'] = df['dislikes'].fillna(0).astype(int)

save_path = "../data/comments_clean.h5"
df.to_hdf(save_path, key='df', mode='w')
