==> notebooks/4_dataset.ipynb <==
# Preparing the dataset

Pulling manual data classification and placing it into train-validate datasets

import pandas as pd

df = pd.read_excel('../data/comments_rlhf.xlsx')
df.head(3)

df_rlhf = df.loc[df['Puzzled'].notna()].copy()

import numpy as np
emotion_labels = ['sadness', 'joy', 'love', 'anger', 'fear', 'surprise', 'Puzzled']
df_rlhf['surprise'] = np.minimum(df_rlhf['surprise'] + df_rlhf['Puzzled'], 1)
df_rlhf['emotion'] = df_rlhf['emotion'].replace({'?puzzled': 'surprise'})

drop_cols = [
    'Unnamed: 0', 
    'Puzzled', 
    'chapter_explanation_update',
    'question_update',
    'question_explanation_update',
    'qc_explanation_update'
]
df_rlhf = df_rlhf.drop(drop_cols, axis=1)

rename_cols = ['Frustration', 'Humour']

df_rlhf = df_rlhf.rename(columns={k: k.lower() for k in rename_cols})
df_rlhf.loc[:, 'theme'] = df_rlhf.loc[:, 'theme'].str.lower()

df_rlhf.head(3)

tone_labels = ['positive', 'negative', 'neutral']
emotion_labels = ['sadness', 'joy', 'love', 'anger', 'fear', 'surprise']
theme_labels = ['educational', 'giving feedback', 'asking a question', 'insulting', 'supporting', 'humour', 'frustration']

df_rlhf[tone_labels + emotion_labels + theme_labels] = df_rlhf[tone_labels + emotion_labels + theme_labels].astype(int)
df_rlhf.dtypes

df_rlhf.loc[:, ['tone', 'emotion', 'theme']].describe()

df_rlhf

rlhf_jsonl = '../data/comments_rlhf.jsonl'
df_rlhf.to_json(rlhf_jsonl, orient="records", lines=True)

from datasets import load_dataset

def train_test_split_ds(ds, test_size=0.2):
    tmp = ds
    ds2 = tmp['train'].train_test_split(test_size=test_size)
    ds3 = ds2['test'].train_test_split(test_size=0.5)
    tmp['train'] = ds2['train']
    tmp['test'] = ds3['train']
    tmp['validate'] = ds3['test']
    return tmp

data_files = {
    'train': rlhf_jsonl
}

ds = load_dataset("json", data_files=data_files)
ds = train_test_split_ds(ds)

ds.push_to_hub('quesmed/comment_sentiment', private=True)
