==> notebooks/4_prodigy.ipynb <==
import pandas as pd

read_path = "../data/comments_clean.h5"
df = pd.read_hdf(read_path, key='df')
df.sample(3)

tone_labels = ['negative', 'neutral', 'positive']
emotion_labels = ['sadness', 'joy', 'love', 'anger', 'fear', 'surprise']
theme_labels = ['clinical update', 'community', 'question', 'education', 'advocating', 'dissuading', 'other']

import json
import os


df['text'] = df['comment']
df['meta'] = ""

df['meta'] = (
  df['id'].apply(lambda x: f'{{"annotated": true, "id": "{x}", ') +
  df['createdAt'].apply(lambda x: f'"createdAt": "{x.strftime("%d/%m/%Y")}", ') +
  df['likes'].apply(lambda x: f'"likes": {x}, ') +
  df['dislikes'].apply(lambda x: f'"dislikes": {x}}}')
)

filename = f'../datasets/comments.jsonl'
df.loc[:, ['text', 'meta']].to_json(filename, orient='records', lines=True)
tmp_filename = '../datasets/tmp.jsonl'

with open(filename, 'r') as f:
  with open(tmp_filename, 'w') as f2:
    for line in f:
      row = json.loads(line)
      try:
        row['meta'] = json.loads(row['meta'])
      except:
        print(row['meta'])
        raise
      f2.write(json.dumps(row, ensure_ascii=False) + "\n")

os.remove(filename)
os.rename(tmp_filename, filename)
