==> notebooks/5_train.ipynb <==
# Load Dataset from HF
Loading from Quesmed organization

from datasets import load_dataset
import os

os.environ['TOKENIZERS_PARALLELISM'] = 'true'

ds = load_dataset('quesmed/comment_sentiment', token=True)
ds

from datasets import Dataset
def isolate_dataset(ds: Dataset, feature: str):
    cols = ds.column_names['train']
    col_keep = {'comment', feature}
    
    ds_filter = ds.remove_columns(col_keep.symmetric_difference(cols))
    ds_filter = ds_filter.rename_column(feature, 'label')
    ds_filter = ds_filter.class_encode_column('label')

    return ds_filter

## Setup model and trainer

from transformers import AutoModelForSequenceClassification
from transformers import AutoTokenizer, AutoConfig
import torch

device = torch.device("mps" if torch.backends.mps.is_available() else "cpu")

def init_model(model_path: str):
    tokenizer = AutoTokenizer.from_pretrained(model_path)
    config = AutoConfig.from_pretrained(model_path)
    model = AutoModelForSequenceClassification.from_pretrained(model_path)

    return (tokenizer, config, model)

def print_summary(result):
    print(f"Time: {result.metrics['train_runtime']:.2f}")
    print(f"Samples/second: {result.metrics['train_samples_per_second']:.2f}")

from transformers import Trainer, TrainingArguments, logging
from datasets import Dataset
import numpy as np
import evaluate

metric = evaluate.load('accuracy')

def compute_metrics(eval_pred):
    logits, labels = eval_pred
    predictions = np.argmax(logits, axis=-1)
    return metric.compute(predictions=predictions, references=labels)

logging.set_verbosity_error()

def setup_trainer(name: str, dataset: Dataset, model, tokenizer, push_to_hub=False):
    model_name = f"../fine-tuning-chkp/{name}"

    training_args = TrainingArguments(
        output_dir=model_name,
        evaluation_strategy="epoch",
        save_strategy="epoch",
        logging_steps=0.2,
        num_train_epochs=5,
        learning_rate=2e-5,
        weight_decay=0.01,
        metric_for_best_model="accuracy",
        load_best_model_at_end=True,
        disable_tqdm=False,
        push_to_hub=push_to_hub,
        hub_model_id=f"quesmed/{name}",
        hub_strategy="end",
        use_mps_device=True
    )

    trainer = Trainer(
        model=model,
        args=training_args,
        train_dataset=dataset['train'],
        eval_dataset=dataset['validate'],
        compute_metrics=compute_metrics,
        tokenizer=tokenizer
    )

    return (trainer, training_args)

# Fine-tuning Tone

tone_tokenizer, tone_config, tone_model = init_model("cardiffnlp/twitter-roberta-base-sentiment-latest")

ds_tone = isolate_dataset(ds, 'tone')

ds_tone = ds_tone.map(
  lambda row: tone_tokenizer(row['comment'], max_length=512, padding='max_length', truncation=True, return_tensors='pt'), 
  batched=True,
  remove_columns=['comment']
)

ds_tone['train'].features

# Hyperparameter search
Using [optuna](https://optuna.org/) to find the optimal training parameters.

tone_trainer, tone_args = setup_trainer('tone', dataset=ds_tone, model=tone_model, tokenizer=tone_tokenizer, push_to_hub=True)

def model_init_tone():
    _, _, model = init_model("cardiffnlp/twitter-roberta-base-sentiment-latest")
    return model

trainer = Trainer(
    model_init=model_init_tone,
    args=tone_args,
    train_dataset=ds_tone["train"],
    eval_dataset=ds_tone["validate"],
    tokenizer=tone_tokenizer,
    compute_metrics=compute_metrics
)

best_run_tone = trainer.hyperparameter_search(n_trials=10, direction="maximize")

best_run_tone

for n, v in best_run_tone.hyperparameters.items():
    setattr(tone_trainer.args, n, v)

result = tone_trainer.train()
print_summary(result)

print(tone_trainer.state.best_model_checkpoint)
tone_trainer.save_model('../fine-tuning-final/tone')


# Emotion fine-tuning

emotion_tokenizer, emotion_config, emotion_model = init_model("bhadresh-savani/distilbert-base-uncased-emotion")


# emotion_labels = ['sadness', 'joy', 'love', 'anger', 'fear', 'surprise']

ds_emotion = isolate_dataset(ds, 'emotion')

ds_emotion = ds_emotion.map(
  lambda row: emotion_tokenizer(row['comment'], max_length=512, padding='max_length', truncation=True, return_tensors='pt'), 
  batched=True,
  remove_columns=['comment']
)

ds_emotion['train'].features

emotion_trainer, emotion_args = setup_trainer('emotion', dataset=ds_emotion, model=emotion_model, tokenizer=emotion_tokenizer, push_to_hub=True)

def model_init_emotion():
    _, _, model = init_model("bhadresh-savani/distilbert-base-uncased-emotion")
    return model

trainer = Trainer(
    model_init=model_init_emotion,
    args=emotion_args,
    train_dataset=ds_emotion["train"],
    eval_dataset=ds_emotion["validate"],
    tokenizer=emotion_tokenizer,
    compute_metrics=compute_metrics
)

best_run_emotion = trainer.hyperparameter_search(n_trials=10, direction="maximize")

best_run_emotion

for n, v in best_run_emotion.hyperparameters.items():
    setattr(emotion_trainer.args, n, v)

result = emotion_trainer.train()
print_summary(result)

result = emotion_trainer.train()
print_summary(result)

print(emotion_trainer.state.best_model_checkpoint)
emotion_trainer.save_model('../fine-tuning-final/emotion')

# Theme fine-tuning

theme_tokenizer, theme_config, theme_model = init_model("facebook/bart-large-mnli")

print(ds['train'][0])

theme_labels = ['educational', 'giving feedback', 'asking a question', 'insulting', 'supporting', 'humour', 'frustration']

template="This example is {}."

for theme in theme_labels:
    def create_theme_ds(row):
      hypothesis = template.format(theme)
      label = 2 if row[theme] == 1 else 0 # convert to 2: entailment for TRUE
      
      return {
         'text': row['text'],
         'hypothesis': hypothesis,
         'label': label
      }

feature = 'theme'
cols = ds.column_names['train']
col_keep = {'text', feature}

ds_theme = ds.remove_columns(col_keep.symmetric_difference(cols))
ds_theme = ds_theme.rename_column(feature, 'label')

from transformers import AutoModelForSequenceClassification, TrainingArguments, Trainer
import random


theme_tokenizer, theme_config, theme_model = init_model("facebook/bart-large-mnli")
# Linear(in_features=1024, out_features=3, bias=True)
# {0: 'contradiction', 1: 'neutral', 2: 'entailment'}

theme_labels = ['clinical update', 'community', 'question', 'education', 'advocating', 'dissuading', 'other']
num_labels = len(theme_labels)
template="This example is {}."

def create_input_sequence(sample):
    text = sample['text']
    label = sample['label'][0]
    contradiction_labels = theme_labels[:]
    label_idx = contradiction_labels.index(label)
    contradiction_labels.pop(label_idx)

    encoded_sequence = theme_tokenizer(
        text,
        [template.format(label)],
        # max_length=512,
        # padding='max_length', 
        truncation=True, 
        return_tensors='pt'
    )
    encoded_sequence['labels'] = [2]
    encoded_sequence['input_sentence'] = theme_tokenizer.batch_decode(encoded_sequence.input_ids)
    return encoded_sequence

ds_theme_encoded = ds_theme.map(
    create_input_sequence, 
    batched=True, 
    batch_size=1,
    remove_columns=["label", "text"]
)

ds_theme_encoded


ds_theme_encoded['train'][0]

premise = ds_theme['train'][0]['text']
template= "This example is {}."
hypothesis = template.format(ds_theme['train'][0]['label'])

# run through model pre-trained on MNLI
x = theme_tokenizer(premise, hypothesis, 
                           truncation_strategy='only_first',
        return_tensors='pt')
x
# logits = theme_model(x.to(device))[0]

# # we throw away "neutral" (dim 1) and take the probability of
# # "entailment" (2) as the probability of the label being true 
# entail_contradiction_logits = logits[:,[0,2]]
# probs = entail_contradiction_logits.softmax(dim=1)
# prob_label_is_true = probs[:,1]
# prob_label_is_true

ds_theme_encoded['train'][0]

theme_trainer, theme_args = setup_trainer('theme', dataset=ds_theme_encoded, model=theme_model, tokenizer=theme_tokenizer)

result = theme_trainer.train()
print_summary(result)

theme_trainer.save_model('../fine-tuning-final/theme')
